chore: reorganise admin e2e by domain + tighten examples/minimal imports
Two small DX cleanups, no logic changes.

examples/minimal
- Switch `auth` import from `@plumix/core/auth` to `plumix` — the
  minimal example is the reference for "what a real consumer's
  package.json looks like", and real consumers install `plumix` +
  their runtime adapter, not core directly. `auth` already re-exports
  through `plumix`'s top-level entry.
- Drop `@plumix/core` from the example's dependencies. Any future
  `import "@plumix/core"` now fails at install/build time — the
  missing dep is a simpler guard than an eslint rule.

Admin e2e layout (grouped by domain, not by screen-split accident)
- bootstrap.spec.ts + login.spec.ts → `auth.spec.ts`
- editor.spec.ts → merged into `content.spec.ts` (list + new + edit
  + meta-box sidebar in one file, in user-journey order)
- users-edit.spec.ts → merged into `users.spec.ts` (list + invite +
  edit + delete + profile in one file)
- `taxonomies.spec.ts` + `dashboard.spec.ts` unchanged (already
  single-file per domain)

No test logic changed — describe blocks moved verbatim, header
comments added to the merged files so the journey ordering is
explicit. Zero duplicate imports / factories / describe wrappers
(verified by find-bugs). All `beforeEach` hooks stay inside their
describes, no cross-describe pollution.

43 workspace tasks + 41 admin e2e + 389 core tests all green.
